See extended description if you want important details
Search function - retrieves the data, but I haven't gotten to do
anything substantial with it yet.

puppies.php - Gets the info from the database and prints correctly;
however, if I remove some of the info and just print out name, sex, and
breed, evertyhing freaks out. I think it's because the classes are all
floated left? I don't really know. I am leaving all the information for
now so it works when we submit something tomorrow.

pupInfo.php - I don't have this working properly. It does get the info
from the database, but we have to set up the links on puppies.php to
send just the puppy_id of the selected puppy so I can query that and
get the correct info. I will try to work on this tomorrow morning I
guess.

css/pupInfo.css - All I changed was the margin-right on the puppic
class so that it would look a little better. Nothing huge.