Trying to set up addToCart.php to see if it helps. See extended description
Theoretically, when you click add to cart, it should send the puppy_id
as a $_GET variable to addToCart, which will put it in a session and
redirect the user to a page or have an alert that says they added the
item to their cart. pupCart.php will be a separate page in itself that
just takes the session and gets the stuff from the database
corresponding to the $_SESSION['cart']. pupCart.php also needs to have
the functionality to remove an item from the session if they want to
delete it from the cart. We won't have an option to update quantity
because no two puppies are the same